+++ SourceCode/Data/OverheadLineFeature.cs
@@ -1,0 +1,10 @@
+﻿namespace ModulesRegistry.Data
+{
+    public enum OverheadLineFeature
+    {
+        No = 0,
+        Removable = 1,
+        Fixed = 2,
+        OnlyPosts = 3
+    }
+}