+++ SourceCode/Data/SignalFeature.cs
@@ -1,0 +1,9 @@
+﻿namespace ModulesRegistry.Data
+{
+    public enum SignalFeature
+    {
+        No = 0,
+        Optional = 1,
+        Fixed = 2
+    }
+}