+++ SourceCode/Data/StationTrack.cs
@@ -11,6 +11,7 @@ namespace ModulesRegistry.Data
         public int StationId { get; set; }
         public string Designation { get; set; }
         public short DisplayOrder { get; set; }
+        public bool IsThroughTrack { get; set; }
         public bool IsSiding { get; set; }
         public bool IsScheduled { get; set; }
         public double MaxTrainLength { get; set; }